OPTIMIZE: split type files
